[feat] 남은 시간 response에 따라 설정 / [refactor] 컴포넌트 분리
