+++ dutchiepay/src/app/_components/_mypage/_modify/ProfileImgButton.jsx
@@ -3,7 +3,7 @@
 import '@/styles/globals.css';
 import '@/styles/mypage.css';
 
-import getImage from '@/app/_util/GetImage';
+import getImage from '@/app/_util/getImage';
 import { useRef } from 'react';
 
 export default function ProfileImgButton({ setModifyInfo }) {
